104 put long back

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,24 +7,24 @@
 
 int main(void)
 {
-	unsigned a, b, c, i;
+	unsigned long a, b, c, i;
 
 	a = 1;
 	b = 2;
 	c = 0;
 
-	printf("%u, %u", a, b);
+	printf("%lu, %lu", a, b);
 
 	for (i = 0; i < 32; i++)
 	{
 		c = a + b;
-		printf(", %u", c);
+		printf(", %lu", c);
 
 		a = b + c;
-		printf(", %u", a);
+		printf(", %lu", a);
 
 		b = c + a;
-		printf(", %u", b);
+		printf(", %lu", b);
 	}
 	putchar('\n');
